+++ app/src/main/java/com/andresDev/puriapp/data/model/PedidoDetallesGeneralesResponse.kt
@@ -27,7 +27,7 @@ data class PedidoDetallesGeneralesResponse(
     data class DetallePedidoDTO(
         val id: Long,
         val producto: ProductoDTO,
-        val cantidad: Int,
+        val cantidad: Double,
         val precioUnitario: Double,
         val subtotal: Double
     ) {
